move:#153 로딩 파일 컴포넌트 폴더로 이동
